poisson_eqn/requirements.txt
jax>=0.4.13
jaxlib>=0.4.13
flax>=0.7.0
optax>=0.1.7
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
hydra-core>=1.3.0
wandb>=0.15.0
ml_collections>=0.1.1
tqdm>=4.65.0
# Add other dependencies as needed, e.g.:
# seaborn>=0.12.0 